fix: remove dependency of calling sku APIs (#62)
We don't need any other skus information because the sku is fixed to the
value presented in the machine.Spec. In the end, we just need the sku
name being passed to the CreateAgentPool API.

With this change, the sku client is never used in gpu-provisioner. We
still keep relevant code (instancetype provide and sku client) for now
and we can do a cleanup later.

---------

Co-authored-by: guofei <guofei@microsoft.com>